full clock semi works

+++ include/Time.hpp
@@ -1,6 +1,6 @@
 struct Time{
-    int year    = 0;
-    int hour    = 0;
-    int minute  = 0;
+    float year    = 0;
+    float hour    = 0;
+    float minute  = 0;
     long second = 0;
 };
